refactor: combine routing types to a single module
