+++ src/lib/api_app/src/api/routes/testbed/mod.rs
@@ -5,9 +5,9 @@ use serde::{Deserialize, Serialize};
 use serde_json::{Value as JSONValue};
 
 use crate::api::{
-    errors::APIRoutingError,
+    routing_types::{APIRoutingError, APIRoutingResponse, ParsedRequest},
     routes::application::get_external_service_bad_response,
-    utils::{get_default_headers, APIRoutingResponse, ParsedRequest},
+    utils::{get_default_headers},
 };
 
 pub mod productizers;
