+++ src/lib/api_app/src/api/utils.rs
@@ -1,48 +1,7 @@
+use super::routing_types::ParsedRequest;
 use http::header::{HeaderMap, HeaderName};
 use http::HeaderValue;
-use http::StatusCode;
-use lambda_http::aws_lambda_events::query_map::QueryMap;
 use lambda_http::{Body, Request, RequestExt};
-use serde_json::json;
-
-use crate::api::errors::APIRoutingError;
-
-#[derive(Debug)]
-pub struct APIRoutingResponse {
-    pub status_code: StatusCode, // http status code, e.g. 200, 404, 500
-    pub body: String,
-    pub headers: HeaderMap,
-}
-
-impl APIRoutingResponse {
-    pub fn new(status_code: StatusCode, body: &str, headers: HeaderMap) -> Self {
-        Self {
-            status_code,
-            body: body.to_string(),
-            headers,
-        }
-    }
-
-    pub fn from_routing_error(error: APIRoutingError) -> Self {
-        Self::new(
-            error.get_status_code(),
-            json!({
-                "message": error.to_string(),
-            })
-            .to_string()
-            .as_ref(),
-            get_default_headers(),
-        )
-    }
-}
-
-pub struct ParsedRequest {
-    pub path: String,
-    pub method: String,
-    pub query: QueryMap,
-    pub headers: HeaderMap,
-    pub body: String,
-}
 
 /**
  * Convert the lambda_http::Request to a parsed_request.
